backend/requirements.txt
Django
djangorestframework
django-cors-headers
gunicorn
psycopg[binary]
whitenoise
dj-database-url